Add files via upload
the project was regarding 2-year sales of the company (1997-1998)) the company has multiple brands and multiple products within it. selling mostly in north America.

The operations performed before the visualizations are data cleaning, data organizing, and creating various measures according to the need. Then, data modelling, and finally, data visualization, with the end product of providing insights.